+++ R/pd_image.R
@@ -37,9 +37,14 @@
 
 pd_image <- function(pd, col = "heat") {
   
-  if (length(pd) != 1) {
+  if (length(pd) > 1) {
     stop("You can only quick image one file at a time.")
   }
+  
+  if (length(pd) == 0) {
+    stop("Provided `pd` has a length of 0 (should be 1).")
+  }
+  
   col <- match.arg(col, c("heat", "redblue"))
 
   pname <- pd_get_name(pd)