+++ NEWS.md
@@ -3,6 +3,7 @@
 * Removed lubridate from imports to fix CRAN note.
 * Removed vignette that depended on too many other libraries; did not want to add them all to Suggests and also relied on external shape file. Will move to somewhere else online. 
 * Removed LazyData and LazyLoad from Description to fix CRAN note.
+* Updated error message in `pd_image()` when input has length of 0.
 
 # prism 0.2.0